use black to format the code

+++ burpui/api/prefs.py
@@ -17,31 +17,34 @@ from ..ext.i18n import LANGUAGES
 from .custom import fields, Resource
 
 bui = current_app  # type: BUIServer
-ns = api.namespace('preferences', 'Preferences methods')
+ns = api.namespace("preferences", "Preferences methods")
 
 
-@ns.route('/ui/hide',
-          endpoint='prefs_ui_hide')
+@ns.route("/ui/hide", endpoint="prefs_ui_hide")
 class PrefsUIHide(Resource):
     """The :class:`burpui.api.prefs.PrefsUI` resource allows you to
     set your UI preferences.
 
     This resource is part of the :mod:`burpui.api.prefs` module.
     """
+
     parser = ns.parser()
-    parser.add_argument('name', dest='client', help='Client to hide')
-    parser.add_argument('agent', dest='server', help='Server to hide')
+    parser.add_argument("name", dest="client", help="Client to hide")
+    parser.add_argument("agent", dest="server", help="Server to hide")
 
-    hidden_model = ns.model('HiddenModel', {
-        'client': fields.String(description='Hidden client name'),
-        'server': fields.String(description='Hidden server name')
-    })
+    hidden_model = ns.model(
+        "HiddenModel",
+        {
+            "client": fields.String(description="Hidden client name"),
+            "server": fields.String(description="Hidden server name"),
+        },
+    )
 
-    @ns.marshal_list_with(hidden_model, description='Success', code=200)
+    @ns.marshal_list_with(hidden_model, description="Success", code=200)
     @ns.doc(
         responses={
-            200: 'Success',
-            403: 'Insufficient permissions',
+            200: "Success",
+            403: "Insufficient permissions",
         },
     )
     def get(self):
@@ -51,19 +54,24 @@ class PrefsUIHide(Resource):
 
         :returns: list
         """
-        if bui.config['WITH_SQL'] and not bui.config['BUI_DEMO'] and not current_user.is_anonymous:
+        if (
+            bui.config["WITH_SQL"]
+            and not bui.config["BUI_DEMO"]
+            and not current_user.is_anonymous
+        ):
             from ..models import Hidden
+
             return Hidden.query.filter_by(user=current_user.name).all()
         return []
 
     @ns.expect(parser)
-    @ns.marshal_with(hidden_model, description='Success', code=200)
+    @ns.marshal_with(hidden_model, description="Success", code=200)
     @ns.doc(
         responses={
-            200: 'Success, object not recorder',
-            201: 'Success, object recorded',
-            403: 'Insufficient permissions',
-            500: 'Internal server error',
+            200: "Success, object not recorder",
+            201: "Success, object recorded",
+            403: "Insufficient permissions",
+            500: "Internal server error",
         },
     )
     def put(self):
@@ -75,12 +83,19 @@ class PrefsUIHide(Resource):
         """
         ret = []
         args = self.parser.parse_args()
-        if bui.config['WITH_SQL'] and not bui.config['BUI_DEMO'] and not current_user.is_anonymous:
+        if (
+            bui.config["WITH_SQL"]
+            and not bui.config["BUI_DEMO"]
+            and not current_user.is_anonymous
+        ):
             from ..ext.sql import db
             from ..models import Hidden
-            client = args.get('client') or None
-            server = args.get('server') or None
-            hidden = Hidden.query.filter_by(client=client, server=server, user=current_user.name).first()
+
+            client = args.get("client") or None
+            server = args.get("server") or None
+            hidden = Hidden.query.filter_by(
+                client=client, server=server, user=current_user.name
+            ).first()
             if not hidden:
                 hide = Hidden(current_user.name, client, server)
                 db.session.add(hide)
@@ -88,7 +103,7 @@ class PrefsUIHide(Resource):
                     db.session.commit()
                 except:  # pragma: no cover
                     db.session.rollback()
-                    self.abort(500, 'Internal server error')
+                    self.abort(500, "Internal server error")
                 return hide, 201
             return hidden
         return ret
@@ -96,9 +111,9 @@ class PrefsUIHide(Resource):
     @ns.expect(parser)
     @ns.doc(
         responses={
-            204: 'Success',
-            403: 'Insufficient permissions',
-            500: 'Internal server error',
+            204: "Success",
+            403: "Insufficient permissions",
+            500: "Internal server error",
         },
     )
     def delete(self):
@@ -107,66 +122,63 @@ class PrefsUIHide(Resource):
         **DELETE** method provided by the webservice.
         """
         args = self.parser.parse_args()
-        if bui.config['WITH_SQL'] and not bui.config['BUI_DEMO'] and not current_user.is_anonymous:
+        if (
+            bui.config["WITH_SQL"]
+            and not bui.config["BUI_DEMO"]
+            and not current_user.is_anonymous
+        ):
             from ..ext.sql import db
             from ..models import Hidden
-            hide = Hidden.query.filter_by(client=(args.get('client') or None), server=(args.get('server') or None), user=current_user.name).first()
+
+            hide = Hidden.query.filter_by(
+                client=(args.get("client") or None),
+                server=(args.get("server") or None),
+                user=current_user.name,
+            ).first()
             if hide:
                 db.session.delete(hide)
                 try:
                     db.session.commit()
                 except:  # pragma: no cover
                     db.session.rollback()
-                    self.abort(500, 'Internal server error')
+                    self.abort(500, "Internal server error")
         return None, 204
 
 
-@ns.route('/ui',
-          endpoint='prefs_ui')
+@ns.route("/ui", endpoint="prefs_ui")
 class PrefsUI(Resource):
     """The :class:`burpui.api.prefs.PrefsUI` resource allows you to
     set your UI preferences.
 
     This resource is part of the :mod:`burpui.api.prefs` module.
     """
+
     parser = ns.parser()
     parser.add_argument(
-        'pageLength',
-        type=int,
-        required=False,
-        help='Number of element per page'
+        "pageLength", type=int, required=False, help="Number of element per page"
     )
     parser.add_argument(
-        'language',
+        "language",
         type=str,
         required=False,
-        help='Language',
-        choices=list(LANGUAGES.keys())
+        help="Language",
+        choices=list(LANGUAGES.keys()),
     )
-    parser.add_argument(
-        'dateFormat',
-        type=str,
-        required=False,
-        help='Date format'
-    )
-    parser.add_argument(
-        'timezone',
-        type=str,
-        required=False,
-        help='Timezone'
-    )
+    parser.add_argument("dateFormat", type=str, required=False, help="Date format")
+    parser.add_argument("timezone", type=str, required=False, help="Timezone")
 
     @staticmethod
     def _user_language(language):
         """Set the current user language"""
         if current_user and not current_user.is_anonymous and language:
-            setattr(current_user, 'language', language)
+            setattr(current_user, "language", language)
 
     def _store_prefs(self, key, val):
         """Store the prefs if persistent storage is enabled"""
-        if bui.config['WITH_SQL'] and not bui.config['BUI_DEMO']:
+        if bui.config["WITH_SQL"] and not bui.config["BUI_DEMO"]:
             from ..ext.sql import db
             from ..models import Pref
+
             pref = Pref.query.filter_by(user=current_user.name, key=key).first()
             if pref:
                 if val:
@@ -187,7 +199,7 @@ class PrefsUI(Resource):
         sess = session._get_current_object()
         ret = {}
         req = MultiDict()
-        for loc in ['values', 'json']:
+        for loc in ["values", "json"]:
             data = getattr(request, loc, None)
             if data:
                 req.update(data)
@@ -196,7 +208,7 @@ class PrefsUI(Resource):
                 continue
             temp = args.get(key)
             if temp:
-                if key == 'language':
+                if key == "language":
                     self._user_language(temp)
                 sess[key] = temp
             elif key in sess:  # pragma: no cover
@@ -208,8 +220,8 @@ class PrefsUI(Resource):
 
     @ns.doc(
         responses={
-            200: 'Success',
-            403: 'Insufficient permissions',
+            200: "Success",
+            403: "Insufficient permissions",
         },
     )
     def get(self):
@@ -229,9 +241,9 @@ class PrefsUI(Resource):
     @ns.expect(parser, validate=True)
     @ns.doc(
         responses={
-            201: 'Success',
-            403: 'Not allowed',
-            400: 'Missing parameters',
+            201: "Success",
+            403: "Not allowed",
+            400: "Missing parameters",
         },
     )
     def put(self):
@@ -241,9 +253,9 @@ class PrefsUI(Resource):
     @ns.expect(parser)
     @ns.doc(
         responses={
-            200: 'Success',
-            403: 'Not allowed',
-            400: 'Missing parameters',
+            200: "Success",
+            403: "Not allowed",
+            400: "Missing parameters",
         },
     )
     def delete(self):
@@ -255,14 +267,12 @@ class PrefsUI(Resource):
             temp = args.get(key)
             if temp:
                 del sess[key]
-                if bui.config['WITH_SQL']:
+                if bui.config["WITH_SQL"]:
                     from ..ext.sql import db
                     from ..models import Pref
+
                     try:
-                        Pref.query.filter_by(
-                            user=current_user.name,
-                            key=key
-                        ).delete()
+                        Pref.query.filter_by(user=current_user.name, key=key).delete()
                         db.session.commit()
                     except:  # pragma: no cover
                         db.session.rollback()
@@ -273,9 +283,9 @@ class PrefsUI(Resource):
     @ns.expect(parser, validate=True)
     @ns.doc(
         responses={
-            200: 'Success',
-            403: 'Not allowed',
-            400: 'Missing parameters',
+            200: "Success",
+            403: "Not allowed",
+            400: "Missing parameters",
         },
     )
     def post(self):
